requirements.txt
Flask==2.3.2
Flask-MySQLdb==2.0.0
python-dotenv==1.0.0
twilio==8.3.0
PyJWT==2.8.0
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
cryptography==41.0.4